+++ core/src/main/java/react4j/react4j.js
@@ -5,8 +5,6 @@
 goog.provide('react4j');
 
 /** @define {string} */
-goog.define('react4j.environment', 'production');
-/** @define {string} */
 goog.define('react4j.enable_component_names', 'false');
 /** @define {string} */
 goog.define('react4j.check_component_state_invariants', 'false');